make video codec switchable by API call

+++ toxav/ring_buffer.h
@@ -33,8 +33,8 @@ extern "C" {
 typedef struct RingBuffer RingBuffer;
 bool rb_full(const RingBuffer *b);
 bool rb_empty(const RingBuffer *b);
-void *rb_write(RingBuffer *b, void *p, uint8_t data_type_);
-bool rb_read(RingBuffer *b, void **p, uint8_t *data_type_);
+void *rb_write(RingBuffer *b, void *p, uint64_t data_type_);
+bool rb_read(RingBuffer *b, void **p, uint64_t *data_type_);
 RingBuffer *rb_new(int size);
 void rb_kill(RingBuffer *b);
 uint16_t rb_size(const RingBuffer *b);
